Add GV2 accept by PNI invite.

+++ libsignal/service/src/main/java/org/whispersystems/signalservice/api/groupsv2/TemporalCredential.java
@@ -8,13 +8,13 @@ public class TemporalCredential {
   private byte[] credential;
 
   @JsonProperty
-  private int redemptionTime;
+  private long redemptionTime;
 
   public byte[] getCredential() {
     return credential;
   }
 
-  public int getRedemptionTime() {
+  public long getRedemptionTime() {
     return redemptionTime;
   }
 }
